+++ CHANGELOG.md
@@ -7,9 +7,11 @@ All notable changes to this project will be documented in this file based on the
 * Removed adapters completely (Repository is now directly using `\Elasticsearch\Client`)
 * Removed `ElasticaQuery`
 * Removed `ruflin/elastica` from list of dependencies
+* Removed method `ElasticsearchRepository::deleteIndex()` in favor of `IndexManager::deleteIndex()`
 ### Bugfixes
 ### Added
 * `postSave` and `postDelete()` hooks for repositories
+* Index management features via `IndexManager`
 ### Improvements
 ### Deprecated